requirements.txt
python-dotenv
asyncpg
aiogram
asyncpg

